feat(vpn): Move worker stuff to dedicated package
It also tightly couple VpnService and VpnWorker as dependency inversion
 was removed.
 As both are supposed to work together, it is fine.
