+++ app/src/main/java/org/adaway/vpn/dns/DnsPacketProxy2.java
@@ -22,7 +22,7 @@ import org.adaway.db.entity.HostEntry;
 import org.adaway.db.entity.ListType;
 import org.adaway.model.vpn.VpnModel;
 import org.adaway.util.AppExecutors;
-import org.adaway.vpn.VpnNetworkException;
+import org.adaway.vpn.worker.VpnNetworkException;
 import org.adaway.vpn.dns.DnsPacketProxy.EventLoop;
 import org.pcap4j.packet.IpPacket;
 import org.pcap4j.packet.IpSelector;
